django_app/requirements.txt
django #new version of django
django-4-jet #beautiful admin panel :)
django-extensions #additional library
djangorestframework #for rest api
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
drf-yasg